requirements.txt
aescipher==6.0.0
amqp==5.3.1
asgiref==3.10.0
attrs==25.4.0
billiard==4.2.2
block-disposable-email==2.0.0
celery==5.5.3
certifi==2025.11.12
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudinary==1.44.1
cryptography==46.0.3
Django==5.2.8
django-celery==3.1.17
django-cors-headers==4.7.0
django-disposable-email-checker==2.0.4
django-redis==6.0.0
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
easyrsa==3.1.0
encryptedsocket==3.0.0
idna==3.11
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.5.4
mailjet-rest==1.5.1
munch==4.0.0
omnitools==0.0.268
packaging==25.0
password-validator==1.0
phonenumbers==9.0.20
pillow==12.0.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pyasynchat==1.0.4
pyasyncore==1.0.4
pycparser==2.23
pycryptodome==3.23.0
pycryptodomex==3.23.0
pyftpdlib==2.1.0
PyJWT==2.10.1
pyOpenSSL==25.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
redis==7.0.1
referencing==0.37.0
requests==2.32.5
rpds-py==0.28.0
six==1.17.0
sqlparse==0.5.3
threadwrapper==0.14.0
typing_extensions==4.15.0
tzdata==2025.2
unencryptedsocket==2.2.0
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
whitenoise==6.10.0
